+++ src/map/_mulle_bigmap_private.h
@@ -12,55 +12,55 @@
 //
 // some somewhat obscure facilities for map
 //
-static inline int  __mulle_keyvaluepair_is_empty( struct _mulle_keyvaluepair *pair, void *not_a_key_marker) mulle_nonnull_first;
-static inline int  __mulle_keyvaluepair_is_empty( struct _mulle_keyvaluepair *pair, void *not_a_key_marker)
+static inline int  __mulle_pointerpair_is_empty( struct mulle_pointerpair *pair, void *not_a_key_marker) mulle_nonnull_first;
+static inline int  __mulle_pointerpair_is_empty( struct mulle_pointerpair *pair, void *not_a_key_marker)
 {
    return( pair->_value == 0 && pair->_key == not_a_key_marker);
 }
 
 
-static inline int  _mulle_keyvaluepair_is_empty( struct _mulle_keyvaluepair *pair,
+static inline int  _mulle_pointerpair_is_empty( struct mulle_pointerpair *pair,
                                                  struct mulle_container_keyvaluecallback *callback) mulle_nonnull_first_second;
-static inline int  _mulle_keyvaluepair_is_empty( struct _mulle_keyvaluepair *pair,
+static inline int  _mulle_pointerpair_is_empty( struct mulle_pointerpair *pair,
                                                  struct mulle_container_keyvaluecallback *callback)
 {
-   return( __mulle_keyvaluepair_is_empty( pair, callback->keycallback.not_a_key_marker));
+   return( __mulle_pointerpair_is_empty( pair, callback->keycallback.not_a_key_marker));
 }
 
 
-static inline int  __mulle_keyvaluepair_is_queue( struct _mulle_keyvaluepair *pair,
+static inline int  __mulle_pointerpair_is_queue( struct mulle_pointerpair *pair,
                                                   void *not_a_key_marker) mulle_nonnull_first;
-static inline int  __mulle_keyvaluepair_is_queue( struct _mulle_keyvaluepair *pair,
+static inline int  __mulle_pointerpair_is_queue( struct mulle_pointerpair *pair,
                                                   void *not_a_key_marker)
 {
    return( pair->_value != 0 && pair->_key == not_a_key_marker);
 }
 
 
-static inline int  _mulle_keyvaluepair_is_queue( struct _mulle_keyvaluepair *pair,
+static inline int  _mulle_pointerpair_is_queue( struct mulle_pointerpair *pair,
                                                  struct mulle_container_keyvaluecallback *callback) mulle_nonnull_first_second;
-static inline int  _mulle_keyvaluepair_is_queue( struct _mulle_keyvaluepair *pair,
+static inline int  _mulle_pointerpair_is_queue( struct mulle_pointerpair *pair,
                                                  struct mulle_container_keyvaluecallback *callback)
 {
-   return( __mulle_keyvaluepair_is_queue( pair, callback->keycallback.not_a_key_marker));
+   return( __mulle_pointerpair_is_queue( pair, callback->keycallback.not_a_key_marker));
 }
 
 
-static inline int  __mulle_keyvaluepair_is_storage( struct _mulle_keyvaluepair *pair,
+static inline int  __mulle_pointerpair_is_storage( struct mulle_pointerpair *pair,
                                                     void *not_a_key_marker) mulle_nonnull_first;
-static inline int  __mulle_keyvaluepair_is_storage( struct _mulle_keyvaluepair *pair,
+static inline int  __mulle_pointerpair_is_storage( struct mulle_pointerpair *pair,
                                                     void *not_a_key_marker)
 {
    return( pair->_key != not_a_key_marker);
 }
 
 
-static inline int  _mulle_keyvaluepair_is_storage( struct _mulle_keyvaluepair *pair,
+static inline int  _mulle_pointerpair_is_storage( struct mulle_pointerpair *pair,
                                                    struct mulle_container_keyvaluecallback *callback) mulle_nonnull_first_second;
-static inline int  _mulle_keyvaluepair_is_storage( struct _mulle_keyvaluepair *pair,
+static inline int  _mulle_pointerpair_is_storage( struct mulle_pointerpair *pair,
                                                    struct mulle_container_keyvaluecallback *callback)
 {
-   return( __mulle_keyvaluepair_is_storage( pair, callback->keycallback.not_a_key_marker));
+   return( __mulle_pointerpair_is_storage( pair, callback->keycallback.not_a_key_marker));
 }
 
 #endif /* _mullebigmap_private_h */