+++ src/mulle_container.h
@@ -34,7 +34,7 @@
 #ifndef mulle_container__h__
 #define mulle_container__h__
 
-#define MULLE_CONTAINER_VERSION  ((0 << 20) | (3 << 8) | 0)
+#define MULLE_CONTAINER_VERSION  ((0 << 20) | (4 << 8) | 0)
 
 #include <mulle_allocator/mulle_allocator.h>
 
@@ -63,11 +63,12 @@
 #include "_mulle_queue.h"
 #include "mulle_pointerset.h"
 #include "mulle_pointerarray.h"
+#include "mulle_pointerpairarray.h"
 
 #include "mulle_container_callback.h"
 #include "mulle_container_operation.h"
 
-#if MULLE_ALLOCATOR_VERSION < ((1 << 20) | (1 << 8) | 0)
+#if MULLE_ALLOCATOR_VERSION < ((1 << 20) | (3 << 8) | 0)
 # error "mulle_allocator is too old"
 #endif