+++ RELEASENOTES.md
@@ -1,3 +1,11 @@
+# 0.4
+
+* upgrade mulle_allocator 
+* rename `_mulle_buffer_add_string_with_length` to 
+  `_mulle_buffer_add_string_with_maxlength` and fix the assert
+* added mulle_pointerpairarray
+* renamed keyvaluepair to pointerpair
+
 # 0.3
 
 * Fix problems with not_a_key_marker being different than 0